Dépôt type période pour les heures additionnelles
